style: 네비게이션바,헤더 컴포넌트 스타일 적용
