+++ package.json
@@ -32,6 +32,7 @@
     "@types/node": "20.3.0",
     "@types/nodemailer": "6.4.8",
     "@types/uuid": "9.0.2",
+    "autoprefixer": "10.4.14",
     "eslint": "8.43.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-typescript": "17.0.0",
@@ -40,6 +41,8 @@
     "eslint-plugin-jsx-a11y": "6.7.1",
     "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "4.6.0",
-    "prettier": "2.8.8"
+    "postcss": "8.4.24",
+    "prettier": "2.8.8",
+    "tailwindcss": "3.3.2"
   }
 }
